Try again: move the test to the server package to make jacaco happy.
